jurassic-journalists/requirements.txt
kivy==1.11.1
Pillow==6.1.0